requirements.txt
python-telegram-bot
google-generativeai
faiss-cpu
sentence-transformers
numpy
python-dotenv
